requirements.txt
gradio==3.41.2
Pillow==9.5.0
numpy==1.24.2
torch==2.0.1
transformers==4.34.0
